Moving towards a VFS
File descriptors are now the sole property of processes, as I've
recently learned that they they really should be.
The `fs` part of things thus deals only in inodes, that are stored in a
tree structure, tracked only by its root. It's not a fancy sorted tree
so it's certainly inefficient when looking up an inode structure from an
inode number but it's very much good enough for now!
That tree is as described in [1]: a t-node points to an i-node which may
contain a list of sub-t-nodes if it's the inode of a directory.
Each directory (starting with the root) is updated once in the tree when
it's first traversed, and then its entries are updated in the tree
directly as needed (file deletion, etc...).
All of the path traversal code is in `fs_open`, which is perhaps
ill-named: it converts a path into its corresponding inode structure in
the tree, if any, and creates it if requested.

[1] https://wiki.osdev.org/Hierarchical_VFS_Theory

+++ kernel/include/kernel/wm.h
@@ -8,7 +8,7 @@
 
 #include <kernel/uapi/uapi_wm.h>
 
-#define WM_NOT_DRAWN  (1 << 31) // Window has _never_ been called wm_render_window
+#define WM_NOT_DRAWN  ((uint32_t) 1 << 31) // Window has _never_ been called wm_render_window
 
 /* ufb: the window's buffer in userspace. Used by the client for drawing
  *  operations. We copy this buffer on request to `kfb`.
